refactor(params): split Params out of axis.cpp into params.cpp
